refactor(gc): split registry and CLI modules
Move the GC registry, reconcile helpers, scanner, and CLI handlers into focused submodules while preserving the existing crate::gc public API.

Refs #332
